requirements.txt
numpy>=1.20.0
requests>=2.25.0
aiohttp>=3.8.0
pytz>=2021.1
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
pandas>=1.3.0
yfinance>=0.2.0
futu-api>=9.6.0